+++ src/paywall-previews/Preview5/styled.ts
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { GiCutDiamond } from 'react-icons/gi';
 import { transparentize } from 'polished';
 
@@ -21,18 +21,12 @@ export const StyledPreviewBox = styled(PreviewBox)<any>`
 type ImageWrapperProps = {
   height: any;
   backgroundImage: any;
-  isRestrictedAsset?: boolean;
 };
 
 export const ImageWrapper = styled(ImageHolder)<ImageWrapperProps>`
   height: ${({ height }) => (height < 75 ? '45%' : '65%')};
   background-image: url(${({ backgroundImage }) => backgroundImage});
   margin: 0;
-  ${({ isRestrictedAsset }) =>
-    isRestrictedAsset &&
-    css`
-      filter: brightness(0.3);
-    `}
 `;
 
 export const Overlay = styled.div`